Add pulumi run command tests
