+++ tests/commands/conftest.py
@@ -1,7 +1,7 @@
 from pytest import fixture
 from typer.testing import CliRunner
 
-from data_safe_haven.config import Config
+from data_safe_haven.config import Config, DSHPulumiConfig
 from data_safe_haven.context import ContextSettings
 
 
@@ -88,3 +88,8 @@ def runner_no_context_file(tmp_path):
 @fixture
 def mock_config_from_remote(mocker, config_sres):
     mocker.patch.object(Config, "from_remote", return_value=config_sres)
+
+
+@fixture
+def mock_pulumi_config_from_remote(mocker, pulumi_config):
+    mocker.patch.object(DSHPulumiConfig, "from_remote", return_value=pulumi_config)
