+++ tests/conftest.py
@@ -2,6 +2,7 @@ from pathlib import Path
 from shutil import which
 from subprocess import run
 
+from pulumi.automation import ProjectSettings
 from pytest import fixture
 
 import data_safe_haven.context.context_settings as context_mod
@@ -18,6 +19,12 @@ from data_safe_haven.config.config import (
 )
 from data_safe_haven.context import Context
 from data_safe_haven.external import AzureApi
+from data_safe_haven.infrastructure import SHMProjectManager
+from data_safe_haven.infrastructure.project_manager import (
+    AzureCliSingleton,
+    ProjectManager,
+    PulumiAccount,
+)
 
 
 @fixture(autouse=True, scope="session")
@@ -237,3 +244,56 @@ def mock_key_vault_key(monkeypatch):
 #         return [MockStorageAccountKey("key1"), MockStorageAccountKey("key2")]
 
 #     monkeypatch.setattr(AzureApi, "get_storage_account_keys", mock_get_storage_account_keys)
+
+
+@fixture
+def mock_azure_cli_confirm(monkeypatch):
+    """Always pass AzureCliSingleton.confirm without attempting login"""
+    monkeypatch.setattr(AzureCliSingleton, "confirm", lambda self: None)  # noqa: ARG005
+
+
+@fixture
+def mock_install_plugins(monkeypatch):
+    """Skip installing Pulumi plugins"""
+    monkeypatch.setattr(
+        ProjectManager, "install_plugins", lambda self: None  # noqa: ARG005
+    )
+
+
+@fixture
+def offline_pulumi_account(monkeypatch, mock_azure_cli_confirm):  # noqa: ARG001
+    """Overwrite PulumiAccount so that it runs locally"""
+    monkeypatch.setattr(
+        PulumiAccount, "env", {"PULUMI_CONFIG_PASSPHRASE": "passphrase"}
+    )
+
+
+@fixture
+def local_project_settings(context_no_secrets, mocker):  # noqa: ARG001
+    """Overwrite adjust project settings to work locally, no secrets"""
+    mocker.patch.object(
+        ProjectManager,
+        "project_settings",
+        ProjectSettings(
+            name="data-safe-haven",
+            runtime="python",
+        ),
+    )
+
+
+@fixture
+def shm_stack_manager(
+    context_no_secrets,
+    config_sres,
+    pulumi_config_no_key,
+    mock_azure_cli_confirm,  # noqa: ARG001
+    mock_install_plugins,  # noqa: ARG001
+    mock_key_vault_key,  # noqa: ARG001
+    offline_pulumi_account,  # noqa: ARG001
+    local_project_settings,  # noqa: ARG001
+):
+    return SHMProjectManager(
+        context=context_no_secrets,
+        config=config_sres,
+        pulumi_config=pulumi_config_no_key,
+    )
